# admin/public/api-docs.yaml
openapi: 3.0.3
info:
  title: Juhe Management Service
  description: OpenAI 兼容 API + 渠道转发 + 提示词管理 + 计费财务
  version: 1.0.0
  contact:
    name: Juhe Studio
servers:
  - url: http://localhost:19528
    description: Local development
paths:
  /v1/models:
    get:
      summary: List available models
      operationId: listModels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelList'
...
  /api/tokens:
    get:
      summary: List API keys
      operationId: listTokens
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      summary: Create API key
      operationId: createToken
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                remain_quota:
                  type: integer
                  description: quota in cents
                unlimited_quota:
                  type: boolean
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: sk-<token> or jwt
  schemas:
    ModelList:
      type: object
      properties:
        object:
          type: string
          example: list
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              object:
                type: string
              created:
                type: integer
              owned_by:
                type: string
...